[style] #153 HomeScreen Composable 분리
